requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-Limiter==3.3.1
psycopg==3.1.18
python-dotenv==1.0.0
PyJWT==2.8.0
email-validator==2.0.0
openpyxl==3.1.2
Werkzeug==2.3.7